# [REF-1584] investigate windows reload performance