introduces an `npm_prefer_offline` config(https://docs.npmjs.com/cli/v6/using-npm/config#prefer-offline) which allows npm to use local cache and only read from the registry when necessary. This is to help reduce the reload time on windows


## Performance Test on windows 11 (VM on Mac M2)
Below gives an overview of install times when `reflex.utils.prerequisites.install_frontend_packages` is called
In this function, We first install base packages using `npm install` and then add tailwind and its dependencies as development dependencies(`npm add -d`) then finally add all detected or installable packages (`npm add`). 

NB: Network connection speed may play a factor in the times but gives a general overview


| `npm install` | `npm add` (Tailwind) | `npm add`(All other packages) | Ran after `reflex init` | Ran after hot reload | with prefer-offline |
|:--------------|:---------------------|------------------------------:|-------------------------|----------------------|---------------------|
| 52s           | 18s                  |                            2m | Yes                     | No                   | No                  |
| 2s            | 5s                   |                           40s | No                      | Yes                  | No                  |                                                  |         |
| 3s            | 2s                   |                            2s | No                      | Yes                  | No                  |
| 2s            | 5s                   |                           45s | No                      | Yes                  | No                  |
| 2m            | 6s                   |                            1m | Yes                     | No                   | No                  |
| 3s            | 3s                   |                           36s | No                      | Yes                  | No                  |
| 2s            | 2s                   |                            3s | No                      | Yes                  | No                  |
| 3s            | 2s                   |                            2s | No                      | Yes                  | No                  |
| 2s            | 3s                   |                           23s | No                      | Yes                  | No                  |
| 2s            | 2s                   |                            2s | No                      | Yes                  | No                  |
| 2s            | 1s                   |                            2s | No                      | Yes                  | No                  |
| 14s           | 26s                  |                            2m | Yes                     | No                   | No                  |
| 2s            | 749ms                |                         961ms | No                      | Yes                  | Yes                 |
| 1s            | 2s                   |                            2s | No                      | Yes                  | Yes                 |
| 2s            | 2s                   |                            2s | No                      | Yes                  | Yes                 |
| 1s            | 1s                   |                            2s | No                      | Yes                  | Yes                 |
| 840ms         | 1s                   |                            1s | No                      | Yes                  | Yes                 |
| 2s            | 1s                   |                            3s | No                      | Yes                  | Yes                 |
| 1s            | 1s                   |                            1s | No                      | Yes                  | Yes                 |
|               |                      |                               |                         |                      |                     |



It can be seen that using `prefer-offline` flag produces consistent minimum start times.